network: splitting VPN Site & VPN Site link
